[dd-trace] moving types out to files so they can be referenced directly
